server/python-worker/python-worker/requirements.txt
celery==5.3.4
python-dotenv==1.0.0
amqp==5.2.0
kombu==5.3.4
pymongo==4.6.0
pika==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
